--- em-app/.cargo/config
@@ -1,2 +1,0 @@
-[build]
-target = "x86_64-fortanix-unknown-sgx"
